+++ commons/test/gas.js
@@ -1,5 +1,6 @@
 const { expect } = require('chai')
-const { gasPriceWithinLimits } = require('..')
+const Web3Utils = require('web3-utils')
+const { gasPriceWithinLimits, normalizeGasPrice } = require('..')
 
 const GAS_PRICE_BOUNDARIES = {
   MIN: 1,
@@ -8,7 +9,6 @@ const GAS_PRICE_BOUNDARIES = {
 
 describe('gas', () => {
   describe('gasPriceWithinLimits', () => {
-
     it('should return gas price if gas price is between boundaries', () => {
       // given
       const minGasPrice = 1
@@ -57,6 +57,66 @@ describe('gas', () => {
 
       // Then
       expect(gasPrice).to.equal(initialGasPrice)
+    })
+
+    describe('normalizeGasPrice', () => {
+      it('should work with oracle gas price in gwei', () => {
+        // Given
+        const oracleGasPrice = 20
+        const factor = 1
+
+        // When
+        const result = normalizeGasPrice(oracleGasPrice, factor).toString()
+
+        // Then
+        expect(result).to.equal('20000000000')
+      })
+      it('should work with oracle gas price not in gwei', () => {
+        // Given
+        const oracleGasPrice = 200
+        const factor = 0.1
+
+        // When
+        const result = normalizeGasPrice(oracleGasPrice, factor).toString()
+
+        // Then
+        expect(result).to.equal('20000000000')
+      })
+      it('should increase gas price value from oracle', () => {
+        // Given
+        const oracleGasPrice = 20
+        const factor = 1.5
+
+        // When
+        const result = normalizeGasPrice(oracleGasPrice, factor).toString()
+
+        // Then
+        expect(result).to.equal('30000000000')
+      })
+      it('should respect gas price max limit', () => {
+        // Given
+        const oracleGasPrice = 200
+        const factor = 4
+        const maxInWei = Web3Utils.toWei(GAS_PRICE_BOUNDARIES.MAX.toString(), 'gwei')
+
+        // When
+        const result = normalizeGasPrice(oracleGasPrice, factor, GAS_PRICE_BOUNDARIES).toString()
+
+        // Then
+        expect(result).to.equal(maxInWei)
+      })
+      it('should respect gas price min limit', () => {
+        // Given
+        const oracleGasPrice = 1
+        const factor = 0.01
+        const minInWei = Web3Utils.toWei(GAS_PRICE_BOUNDARIES.MIN.toString(), 'gwei')
+
+        // When
+        const result = normalizeGasPrice(oracleGasPrice, factor, GAS_PRICE_BOUNDARIES).toString()
+
+        // Then
+        expect(result).to.equal(minInWei)
+      })
     })
   })
 })
